test(gf): extract Jest matchers to separate library
