+++ libs/gf/test-util-matchers/src/test-setup.ts
@@ -15,9 +15,3 @@
  * limitations under the License.
  */
 import 'jest-preset-angular/setup-jest';
-
-import { addNgModuleMatchers } from '@energinet-datahub/gf/test-util-matchers';
-import { setUpTestbed } from '@energinet-datahub/gf/test-util-staging';
-
-setUpTestbed();
-addNgModuleMatchers();
