+++ libs/gf/test-util-staging/src/index.ts
@@ -20,8 +20,5 @@ export * from './lib/set-up-angular-testing-library';
 // Angular testbed
 export * from './lib/set-up-testbed';
 
-// Jest matchers
-export * from './lib/add-ngmodule-matchers';
-
 // ng-mocks
 export * from './lib/set-up-ng-mocks';
