+++ libs/gf/util-browser/src/test-setup.ts
@@ -16,6 +16,6 @@
  */
 import 'jest-preset-angular/setup-jest';
 
-import { addNgModuleMatchers } from '@energinet-datahub/gf/test-util-staging';
+import { addNgModuleMatchers } from '@energinet-datahub/gf/test-util-matchers';
 
 addNgModuleMatchers();
